Move createFlatChildArgs() to detach.ts, add version to list, ensure process.cmd matches.
